backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
chromadb==0.4.22
anthropic==0.18.1
langchain==0.1.5
langchain-community==0.0.17
pypdf==3.17.4
sentence-transformers==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
